Update common.jl
A constructor for ExEdge that takes an additional parameter is a useful extension that makes a lot of sense because the ExEdge typed is implemented *specifically* to allow the inclusion of AttritubeDict (s) in addition to the source and target of the edge. A complementary change I'm submitting in the corresponding file is an version of the add_edge! function that makes use of this.